Factor out dev-mode localConfig/migMeta into a helper
